+++ src/math/vector.rs
@@ -6,13 +6,13 @@ use std::ops::{ Add, Div, Mul, Neg, Sub };
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Vector {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 impl Vector {
-    pub fn new(x: f32, y: f32, z: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
         Self { x, y, z }
     }
 
@@ -20,11 +20,11 @@ impl Vector {
         Self::new(0.0, 0.0, 0.0)
     }
 
-    pub fn norm(&self) -> f32 {
+    pub fn norm(&self) -> f64 {
         self.norm_sqr().sqrt()
     }
 
-    pub fn norm_sqr(&self) -> f32 {
+    pub fn norm_sqr(&self) -> f64 {
         self.x * self.x + self.y * self.y + self.z * self.z
     }
 
@@ -34,7 +34,7 @@ impl Vector {
         Self::new(self.x / n, self.y / n, self.z / n)
     }
 
-    pub fn dot(self, rhs: Self) -> f32 {
+    pub fn dot(self, rhs: Self) -> f64 {
         self.x * rhs.x + self.y * rhs.y + self.z * rhs.z
     }
 
@@ -59,10 +59,10 @@ impl Add for Vector {
     }
 }
 
-impl Div<f32> for Vector {
+impl Div<f64> for Vector {
     type Output = Self;
 
-    fn div(self, rhs: f32) -> Self::Output {
+    fn div(self, rhs: f64) -> Self::Output {
         Self::Output::new(
             self.x / rhs,
             self.y / rhs,
@@ -71,10 +71,10 @@ impl Div<f32> for Vector {
     }
 }
 
-impl Mul<f32> for Vector {
+impl Mul<f64> for Vector {
     type Output = Self;
 
-    fn mul(self, rhs: f32) -> Self::Output {
+    fn mul(self, rhs: f64) -> Self::Output {
         Self::Output::new(
             rhs * self.x,
             rhs * self.y,
